Better separation of new JSON and XML serializers, so jackson XML specific annotations are never loaded when using JSON codecs. This is required, as all jackson XML dependencies are marked as provided in gradle build, and are not included by default.
