+++ eureka-client/src/main/java/com/netflix/discovery/converters/jackson/AbstractEurekaJacksonCodec.java
@@ -15,6 +15,7 @@ import com.fasterxml.jackson.databind.ser.PropertyWriter;
 import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
 import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
 import com.netflix.appinfo.InstanceInfo;
+import com.netflix.discovery.converters.jackson.mixin.MiniInstanceInfoMixIn;
 
 /**
  * @author Tomasz Bak
