+++ eureka-client/src/main/java/com/netflix/discovery/converters/jackson/EurekaJsonJacksonCodec.java
@@ -37,8 +37,8 @@ public class EurekaJsonJacksonCodec extends AbstractEurekaJacksonCodec {
     public EurekaJsonJacksonCodec(final KeyFormatter keyFormatter, boolean compact) {
         // JSON
         SimpleModule jsonModule = new SimpleModule();
-        jsonModule.setSerializerModifier(EurekaJacksonModifiers.createJsonSerializerModifier(keyFormatter));
-        jsonModule.setDeserializerModifier(EurekaJacksonModifiers.createJsonDeserializerModifier(keyFormatter, compact));
+        jsonModule.setSerializerModifier(EurekaJacksonJsonModifiers.createJsonSerializerModifier(keyFormatter));
+        jsonModule.setDeserializerModifier(EurekaJacksonJsonModifiers.createJsonDeserializerModifier(keyFormatter, compact));
         jsonMapper.registerModule(jsonModule);
         jsonMapper.setSerializationInclusion(Include.NON_NULL);
         jsonMapper.configure(SerializationFeature.WRAP_ROOT_VALUE, true);
